requirements.txt
Django==5.0.1
djangorestframework==3.14.0
firebase-admin==6.4.0
python-dateutil==2.8.2
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-rest-swagger==2.2.0
drf-yasg==1.21.7